requirements.txt
Flask==2.3.3
markdown2==2.4.10
python-frontmatter==1.0.0 